+++ src/MongoDB.Bson/Serialization/Serializers/SByteSerializer.cs
@@ -102,13 +102,13 @@ namespace MongoDB.Bson.Serialization.Serializers
                 case BsonType.Int32:
                     var int32Value = bsonReader.ReadInt32();
                     value = (sbyte)int32Value;
-                    lostData = (int)value != int32Value;
+                    lostData = value != int32Value;
                     break;
 
                 case BsonType.Int64:
                     var int64Value = bsonReader.ReadInt64();
                     value = (sbyte)int64Value;
-                    lostData = (int)value != int64Value;
+                    lostData = value != int64Value;
                     break;
 
                 case BsonType.String:
